requirements.txt
ragie
streamlit
python-dotenv